pyproject.toml
[project]
name = "schule-ein-interface-geben"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dlt[duckdb]>=1.13.0",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "lib",
    "dspy>=2.6.27",
    "python-dotenv>=1.1.1",
    "polars>=1.31.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "markdown>=3.8.2",
    "markdownify>=1.1.0",
    "selenium>=4.34.2",
    "anytree>=2.13.0",
    "rich>=14.0.0",
    "pytest>=8.4.1",
    "boto3>=1.39.12",
    "poldantic>=0.2.0",
]

[tool.uv.sources]
lib = { workspace = true }

[tool.uv.workspace]
members = ["lib"]
